+++ lib/models/country_config.dart
@@ -4,8 +4,12 @@ class CountryConfig {
   bool noFlag;
   TextStyle textStyle;
   Decoration decoration;
+  bool flatFlag;
+  double flagSize;
   CountryConfig({
     this.noFlag = false,
+    this.flagSize = 30,
+    this.flatFlag = false,
     Decoration? decoration,
     this.textStyle = const TextStyle(
         color: Colors.black, fontSize: 16, fontWeight: FontWeight.w600),